fix(mcp): derive ratio-semantic size from option sets; memoize vendor key probe; tool catalog extraction
