+++ electron/catalog/archetypeWireDefaults.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { ARCHETYPE_WIRE_DEFAULTS } from "./archetypeWireDefaults.generated";
+import { ARCHETYPE_WIRE_DEFAULTS, ARCHETYPE_SIZE_RATIO_SEMANTIC } from "./archetypeWireDefaults.generated";
 
 // 锁住档案默认桥接的关键不变量（生成数据由 check:archetype-defaults 门保证与档案同步）。
 describe("ARCHETYPE_WIRE_DEFAULTS（headless 缺参兜底桥接）", () => {
@@ -27,5 +27,23 @@ describe("ARCHETYPE_WIRE_DEFAULTS（headless 缺参兜底桥接）", () => {
     expect(ARCHETYPE_WIRE_DEFAULTS["volcengine-seedream"].text_to_image["*"].size).toBe("2048x2048");
     expect(ARCHETYPE_WIRE_DEFAULTS["nomi-audio"].text_to_audio["*"].model).toBeTruthy();
     expect(ARCHETYPE_WIRE_DEFAULTS["volcengine-doubao-tts"].text_to_audio["*"].voice).toBeTruthy();
+  });
+});
+
+// size 键比例语义标记（从档案 size 控件选项集 derive，供 headless size 别名闸判「调用方比例能否落到 size」）。
+describe("ARCHETYPE_SIZE_RATIO_SEMANTIC（size 键比例语义桥接）", () => {
+  it("比例族默认（adaptive）但选项集含真比例档 → 记为比例语义（seedance-2.5-apimart 修复点）", () => {
+    // size 默认 "adaptive"（不匹配 /^\d+:\d+$/），旧「按默认值字面猜」会误判像素语义、吞掉调用方比例。
+    expect(ARCHETYPE_SIZE_RATIO_SEMANTIC["seedance-2.5-apimart"].text_to_video).toBe(true);
+    expect(ARCHETYPE_SIZE_RATIO_SEMANTIC["seedance-2.5-apimart"].image_to_video).toBe(true);
+  });
+
+  it("默认值本就是比例形的档案照样登记（apimart seedream）", () => {
+    expect(ARCHETYPE_SIZE_RATIO_SEMANTIC["seedream"].text_to_image).toBe(true);
+  });
+
+  it("像素语义档案不登记（volcengine-seedream / modelscope-image 的 size 选项全是像素）→ 闸会剥掉调用方比例", () => {
+    expect(ARCHETYPE_SIZE_RATIO_SEMANTIC["volcengine-seedream"]).toBeUndefined();
+    expect(ARCHETYPE_SIZE_RATIO_SEMANTIC["modelscope-image"]).toBeUndefined();
   });
 });
