+++ electron/catalog/taskParams.ts
@@ -6,7 +6,7 @@
 // params 的坑，都只在"真实参数构建"里暴露，埋在 2500 行 runtime 里既测不到也容易回归。
 import { firstString, type JsonRecord } from "../jsonUtils";
 import { referenceInputParams } from "./archetypeInput";
-import { ARCHETYPE_WIRE_DEFAULTS } from "./archetypeWireDefaults.generated";
+import { ARCHETYPE_WIRE_DEFAULTS, ARCHETYPE_SIZE_RATIO_SEMANTIC } from "./archetypeWireDefaults.generated";
 import { bodyReferencedParamKeys } from "./paramTranslate";
 import { bodyReferenceSupport, classifyReferenceKey, type ReferenceFamily } from "./referenceReachability";
 
@@ -57,14 +57,26 @@ const RATIO_VALUE_RE = /^\d+\s*:\s*\d+$/;
  * modelscope / agnes / rh-qwen / rh-sora 把它当**像素**读（默认 "2048x2048" / "1024x1024" / "720x1280"）。
  * buildGenerateParams 无差别把调用方比例铺进 size 别名，caller-wins 于是会把像素档案的 size 覆写成 "16:9"，
  * 渲染进 wire body 就是废请求（火山 seedream 直接坏）。故在**看得见所选模式真实默认**的这道缝（extras 与
- * mapping/档案默认在此汇合）加闸：只有当该模式给 `size` 的默认本身是比例形（/^\d+:\d+$/）时，才准调用方比例落到
- * size；默认是像素形或压根没有 size 默认 → 不落（size 那半留给档案像素默认，调用方比例对这个只会说像素的目标不适用）。
- * 语义无歧义的 aspect_ratio / aspectRatio 别名不受此限，照常保留（模板没引用就自然被丢弃）。**纯 derive，无 vendor 名单。**
+ * mapping/档案默认在此汇合）加闸：size 键是比例语义时才准调用方比例落到 size；是像素语义或压根没有 size →
+ * 不落（size 那半留给档案像素默认，调用方比例对这个只会说像素的目标不适用）。
+ *
+ * **判据从档案 size 控件的选项集 DERIVE**（生成期算好、桥进 ARCHETYPE_SIZE_RATIO_SEMANTIC）：选项集里有真比例档
+ * （16:9…）→ 这个 size 键是比例语义，哪怕它的默认值是 "adaptive" 这种比例族自动档（seedance-2.5-apimart t2v
+ * 正是此例：size 默认 "adaptive"、body 只读 size，旧的「按默认值字面 /^\d+:\d+$/ 猜」会误判成像素语义、把调用方
+ * 16:9 剥掉 → 画幅被吞）。只有档案里查不到该 (archetypeId, taskKind)（如自定义/未桥接的档案）才回退到旧的
+ * **默认值字面形状**正则兜底。语义无歧义的 aspect_ratio / aspectRatio 别名不受此限，照常保留（模板没引用就自然被丢弃）。
+ * **纯 derive，无 vendor 名单。**
  */
 function sizeDefaultIsRatioSemantic(
+  archetypeId: string | undefined,
+  taskKind: string,
   archetypeDefaults: Record<string, unknown> | undefined,
   mappingDefaults: Record<string, unknown> | undefined,
 ): boolean {
+  // ① 首选：档案 size 控件选项集 derive 出的比例语义标记（覆盖 "adaptive" 这类默认值猜不出的比例族档）。
+  const emitted = archetypeId ? ARCHETYPE_SIZE_RATIO_SEMANTIC[archetypeId]?.[taskKind] : undefined;
+  if (typeof emitted === "boolean") return emitted;
+  // ② 回退（未桥接的档案）：按合并后真正生效的 size 默认值字面形状猜。
   // 有效默认 = 合并后真正生效的那个（mappingDefaults 是更贴近的兜底、后铺，故它有 size 时以它为准）。
   const effective = mappingDefaults && "size" in mappingDefaults
     ? mappingDefaults.size
@@ -92,7 +104,7 @@ export function applyHeadlessParamDefaults(
   // extras.size 是比例形（调用方 aspect_ratio 铺来的）、但本模式的 size 是像素语义 → 剥掉，让档案像素默认接管。
   // 只针对「比例形的 caller size」，UI 路自己填的真实像素 size 不受影响（它不匹配 RATIO_VALUE_RE）。
   const guarded = extras && typeof extras.size === "string" && RATIO_VALUE_RE.test(extras.size.trim())
-    && !sizeDefaultIsRatioSemantic(archetypeDefaults, mappingDefaults)
+    && !sizeDefaultIsRatioSemantic(archetypeId, taskKind, archetypeDefaults, mappingDefaults)
     ? (() => { const { size: _dropped, ...rest } = extras; return rest; })()
     : extras;
   return applyWireDefaults(applyWireDefaults(guarded, archetypeDefaults), mappingDefaults);
