+++ scripts/gen-archetype-wire-defaults.ts
@@ -14,6 +14,11 @@ type ModeManifest = Record<string, {
   defaultModeId: string;
   modes: Record<string, string>;
 }>;
+// { archetypeId: { taskKind: true } } —— 该模式的 `size` 控件是「比例语义」（选项集里有 \d+:\d+ 档）。
+type SizeRatioSemantic = Record<string, Record<string, boolean>>;
+
+// 一个 size 控件的选项集里出现任意 "16:9" 这样的比例档 → 判定 size 键是比例语义（而非像素语义）。
+const RATIO_OPTION_RE = /^\d+\s*:\s*\d+$/;
 
 // 结构：{ archetypeId: { taskKind: { "*"|vendorKey: { param: default } } } }。
 // 同一档案同一 taskKind 在不同 vendor 下参数枚举不同（vendorParams，B 分层，如 apimart Kling duration=number
@@ -45,6 +50,42 @@ export function buildArchetypeWireDefaults(): WireDefaults {
         out[arch.id][taskKind] = out[arch.id][taskKind] ?? {};
         (out[arch.id][taskKind] as Record<string, Record<string, unknown>>)[vendorKey] = d;
       }
+    }
+  }
+  return out;
+}
+
+/**
+ * 每个模式（taskKind）的 `size` 键到底是**比例语义**还是**像素语义**——从控件选项集 DERIVE，而不是从
+ * 默认值的字面形状猜。
+ *
+ * 为什么必须从这里 derive：headless size 别名闸（taskParams.sizeDefaultIsRatioSemantic）此前只按「默认值
+ * 长得像不像 \d+:\d+」判——但 seedance-2.5-apimart 的 size 默认是 "adaptive"（比例族里的自动档，不匹配那个
+ * 正则），而它的 size 控件选项集正是 ["adaptive","16:9","9:16",...]，且 t2v body 只读 `size`。于是调用方
+ * aspect_ratio="16:9" 铺进的 size 被误判成「像素语义」而剥掉，默认 "adaptive" 胜出，画幅被吞。选项集里只要
+ * 出现一个真比例档（16:9…），这个 size 键就是比例语义——用它填调用方比例是对的。此判据在生成期（能 import
+ * src/config 档案、看得见控件 options）算好，桥过去给 electron 的闸读；闸对没登记的档案回退旧的值形状正则。
+ *
+ * 结构 { archetypeId: { taskKind: true } }：只登记「是比例语义」的（true），像素语义/无 size 控件的不落键
+ * （闸里 `?? 回退` 即可），生成文件更小。同一 taskKind 跨 vendorParams 多套 params 时，任一套的 size 控件
+ * 是比例语义即记 true（调用方比例对该模式适用）。
+ */
+export function buildArchetypeSizeRatioSemantic(): SizeRatioSemantic {
+  const out: SizeRatioSemantic = {};
+  for (const arch of MODEL_ARCHETYPES) {
+    for (const mode of arch.modes) {
+      const taskKind = mode.transportTaskKind ?? arch.transportTaskKind;
+      if (!taskKind) continue;
+      // 通用 params 与每个 vendorParams 覆盖里的 size 控件都算；任一套判为比例语义就登记。
+      const paramSets: Array<typeof mode.params> = [mode.params ?? [], ...Object.values(mode.vendorParams ?? {})];
+      const ratioSemantic = paramSets.some((params) =>
+        (params ?? []).some(
+          (p) => p.key === "size" && (p.options ?? []).some((o) => RATIO_OPTION_RE.test(String(o.value).trim())),
+        ),
+      );
+      if (!ratioSemantic) continue;
+      out[arch.id] = out[arch.id] ?? {};
+      out[arch.id][taskKind] = true;
     }
   }
   return out;
@@ -89,10 +130,13 @@ export function buildArchetypeModeManifest(): ModeManifest {
 
 const BANNER = "// @generated by scripts/gen-archetype-wire-defaults.ts — 请勿手改（改档案后跑 pnpm gen:archetype-defaults）。\n";
 
-export function renderGeneratedFile(defaults: WireDefaults): string {
+export function renderGeneratedFile(defaults: WireDefaults, sizeRatioSemantic: SizeRatioSemantic): string {
   return (
     `${BANNER}// 档案参数默认值桥接：headless/MCP 生成缺参时 runtime 按 (archetypeId, taskKind) 兜底填，既有值优先。\n` +
-    `export const ARCHETYPE_WIRE_DEFAULTS: Record<string, Record<string, Record<string, Record<string, unknown>>>> = ${JSON.stringify(defaults, null, 2)};\n`
+    `export const ARCHETYPE_WIRE_DEFAULTS: Record<string, Record<string, Record<string, Record<string, unknown>>>> = ${JSON.stringify(defaults, null, 2)};\n` +
+    `\n// size 键比例语义桥接（从档案 size 控件选项集 derive）：headless size 别名闸据此判「调用方比例能否落到 size」，\n` +
+    `// 不再只按默认值字面形状猜（修 seedance-2.5-apimart t2v 的 size 默认 "adaptive" 被误判像素语义、吞掉调用方比例）。\n` +
+    `export const ARCHETYPE_SIZE_RATIO_SEMANTIC: Record<string, Record<string, boolean>> = ${JSON.stringify(sizeRatioSemantic, null, 2)};\n`
   );
 }
 
@@ -116,7 +160,7 @@ const isMain = process.argv[1] === fileURLToPath(import.meta.url);
 if (isMain) {
   const dir = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "../electron/catalog");
   const outputs: Array<{ file: string; content: string }> = [
-    { file: "archetypeWireDefaults.generated.ts", content: renderGeneratedFile(buildArchetypeWireDefaults()) },
+    { file: "archetypeWireDefaults.generated.ts", content: renderGeneratedFile(buildArchetypeWireDefaults(), buildArchetypeSizeRatioSemantic()) },
     { file: "archetypeIdentifiers.generated.ts", content: renderIdentifiersFile(buildArchetypeIdentifierPatterns()) },
     { file: "archetypeModes.generated.ts", content: renderModesFile(buildArchetypeModeManifest()) },
   ];
